cleaned up and organized files
