+++ hrsTree_scrape.py
@@ -137,7 +137,7 @@ def scrapeSectionNames(url):
 
         # Looks for statute code in Regex ex. 123-45.5
         rgx_code = re.search(
-            '(\d+|\w+)\-((\d+\.\d+\w+)|(\d+\.\d+)|(\d+\w+)|(\d+))',
+            '(\d+|\w+)\-((\d+\.\d+\w+)|(\d+\.\d+)|(\d+\w{1})|(\d+))',
             clean_line)
 
         if rgx_code is not None:
@@ -150,6 +150,21 @@ def scrapeSectionNames(url):
             curr_section_name = clean_line.replace(
                 rgx_code.group(0), '').strip()
             curr_chapter_section = rgx_code.group(0).split('-')
+
+            # If the section name begins with /\d+:/ delete it
+            secNameBegin1 = re.search('(^\d+: )', curr_section_name)
+            if secNameBegin1 is not None:
+                curr_section_name = curr_section_name.replace(secNameBegin1.group(0), "")
+
+            # If the curr_chapter_section ends w/ a capital letter
+            # and curr_section_name starts with a lowercase letter
+            # and has stuff after it
+            chapSecEnd = re.search('[A-Z]$', curr_chapter_section[1])
+            secNameBegin = re.search('^[a-z].{3,}', curr_section_name)
+
+            if chapSecEnd is not None and secNameBegin is not None:
+                curr_section_name = curr_chapter_section[1][-1] + curr_section_name
+                curr_chapter_section[1] = curr_chapter_section[1][:-1]
 
             # Check if there are multiple statutes in a line
             found_multiples = checkMultiples(
@@ -175,21 +190,97 @@ def scrapeSectionNames(url):
     return Sections
 
 
+def checkLine(currentLine):
+    retVal = True
+    blackList = ["Chapter", "Subtitle"]
+
+    for word in blackList:
+        if word in currentLine:
+            retVal = False
+
+    return retVal
+
+
 def main():
-    Chapter = {"number": 205, "name": "Land Use Comission", "repealed": False}
+    baseURL = 'http://www.capitol.hawaii.gov/docs/HRS.htm'
+    htmlToParse = requests.get(baseURL)
+    soup = bs(htmlToParse.text, 'lxml')
 
-    Sections = scrapeSectionNames(
-        'http://www.capitol.hawaii.gov/hrs2016/Vol06_Ch0321-0344/HRS0321/HRS_0321-.htm')
+    Divisions = []
+    Titles = []
+    Chapters = []
 
-    if Sections is None:
-        Chapter['repealed'] = True
-    else:
-        Chapter["sections"] = Sections
+    currentDivision = {}
+    currentTitle = {}
+    currentChapter = {}
+    chapterTrigger = 0
 
-    outfile = open('chapter_example.json', 'w')
-    json.dump(Chapter, outfile, sort_keys=True,
+    contents = soup.find_all('p', {'class': 'MsoNormal'})
+
+    for content in contents:
+        currentLine = content.get_text().replace('\r\n ', "")
+        if checkLine(currentLine):
+            if "DIVISION" in currentLine:
+                if len(Titles) > 0:
+                    currentTitle["chapters"] = Chapters
+                    Titles.append(currentTitle)
+
+                    currentDivision["titles"] = Titles
+                    Divisions.append(currentDivision)
+
+                    currentDivision = {}
+                    currentTitle = {}
+                    currentChapter = {}
+                    Titles = []
+                    Chapters = []
+
+                    currentDivision["name"] = currentLine
+                else:
+                    currentDivision["name"] = currentLine
+
+            elif "TITLE" in currentLine:
+                if len(Chapters) > 0:
+                    currentTitle["chapters"] = Chapters
+                    Titles.append(currentTitle)
+
+                    currentTitle = {}
+                    currentChapter = {}
+                    Chapters = []
+
+                    currentTitle["name"] = currentLine
+                else:
+                    currentTitle["name"] = currentLine
+            else:
+                if chapterTrigger == 0:
+                    currentChapter["number"] = currentLine
+                    currentChapter["text"] = ""
+                    chapterTrigger = 1
+
+                elif chapterTrigger == 1:
+                    currentChapter["text"] = currentLine
+                    chapterUrl = content.a['href']
+                    Sections = scrapeSectionNames(chapterUrl)
+
+                    if Sections is not None:
+                        currentChapter['repealed'] = False
+                        currentChapter['sections'] = Sections
+                    else:
+                        currentChapter['repealed'] = True
+
+                    Chapters.append(currentChapter)
+                    currentChapter = {}
+                    chapterTrigger = 0
+
+    currentTitle["chapters"] = Chapters
+    Titles.append(currentTitle)
+
+    currentDivision["titles"] = Titles
+    Divisions.append(currentDivision)
+
+    outfile = open('output/hrsTree.json', 'w')
+    json.dump(Divisions, outfile, sort_keys=True,
               indent=4, separators=(',', ': '))
-    print("Data scraped into chapter_example.json")
+    print("Data scraped into output/hrsTree.json")
 
 
 if __name__ == '__main__':
